+++ lib/volley_motion/error_listener.rb
@@ -1,23 +1,22 @@
 module VolleyMotion
   class ErrorListener
 
-    class << self
-      def build
-        ErrorListener.new
-        # error_listener = Com::Android::Volley::Response::ErrorListener.new
-      end
-    end
+    attr_accessor :activity
 
-    def initialize
+    def initialize(activity)
+      @activity = activity
     end
 
     def onErrorResponse(error)
-      if error.toString == "com.android.volley.AuthFailureError"
-        p "Authentication error"
-      else
-        p error.toString
+      error_result = { :error => :network_error, :message => "Network error."}
+      case error.class.toString
+      when "class com.android.volley.AuthFailureError"
+        error_result = { :error => :auth_failure, :message => "Authentication Failure."}
+      when "class com.android.volley.NoConnectionError"
+        error_result = { :error => :no_connection, :message => "No connection."}
       end
-
+      @activity.error_response(error_result)
     end
+
   end
 end