+++ lib/volley_motion/json_array_auth_request.rb
@@ -1,0 +1,27 @@
+module VolleyMotion
+  class JsonArrayAuthRequest < Com::Android::Volley::Toolbox::JsonArrayRequest
+    class << self
+      def build(url, success_listener, error_listener)
+        array_request = Com::Android::Volley::Toolbox::JsonArrayRequest.new(url, success_listener, error_listener)
+      end
+    end
+
+    def username=(username)
+      @username = username
+    end
+
+    def password=(password)
+      @password = password
+    end
+
+    def getHeaders
+      headers = {}
+      auth_string = @username + ":" + @password
+      auth = "Basic " + convertTo64(auth_string)
+      auth = Java::Lang::String.new(auth)
+      headers[Java::Lang::String.new("Authorization")] = auth
+      headers
+    end
+
+  end
+end