+++ lib/volley_motion/json_array_request.rb
@@ -1,5 +1,5 @@
 module VolleyMotion
-  class JsonArrayRequest
+  class JsonArrayRequest < Com::Android::Volley::Toolbox::JsonArrayRequest
     class << self
       def build(url, success_listener, error_listener)
         array_request = Com::Android::Volley::Toolbox::JsonArrayRequest.new(url, success_listener, error_listener)