+++ lib/volley_motion/json_object_request.rb
@@ -1,0 +1,9 @@
+module VolleyMotion
+  class JsonObjectRequest < Com::Android::Volley::Toolbox::JsonObjectRequest
+    class << self
+      def build(url, success_listener, error_listener)
+        array_request = Com::Android::Volley::Toolbox::JsonObjectRequest.new(url, nil, success_listener, error_listener)
+      end
+    end
+  end
+end