+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Change Log
 ==================
 
-0.0.7 - Add POST request class
+## 0.0.7 - [Unreleased]
+### Added
+* POST request class
+* POST BasicAuth class
+* GET BasicAuth class
+### Changed
+* Change the way that the responders work, now calls on_response on the responder
+* Improved error handling responses
+* Upgraded the Volley library JAR to 1.0.15
 
-0.0.6 - Add work around for gem issue with RubyMotion.
+## 0.0.6
+*  Add work around for gem issue with RubyMotion.